Chromosome Data

Data in this example refer to 337 observations on the secondary association of chromosomes in Brassika; n , which is now the number of chromosomes, equals 3 and X is the number of pairs of bivalents showing association.

Usage

Chromosome_data

Format

A data frame with 2 columns and 4 rows

No.of.Asso

No of Associations

fre

Observed frequencies

Source

Extracted from

Paul, S.R., 1985. A three-parameter generalization of the binomial distribution. Communications in Statistics - Theory and Methods, 14(6), pp.1497-1506.

Available at: doi:10.1080/03610928508828990

Examples

Chromosome_data$No.of.Asso          #extracting the binomial random variables
#> [1] 0 1 2 3
sum(Chromosome_data$fre)            #summing all the frequencies
#> [1] 337