Estimating the shape parameters a and b for Beta-Binomial Distribution

The functions will estimate the shape parameters using the maximum log likelihood method and moment generating function method for the Beta-Binomial distribution when the binomial random variables and corresponding frequencies are given.

EstMGFBetaBin(x,freq)

Arguments

x
vector of binomial random variables.

freq
vector of frequencies.

Value

The output of EstMGFBetaBin will produce the class mgf format consisting

a shape parameter of beta distribution representing for alpha

b shape parameter of beta distribution representing for beta

min Negative loglikelihood value

AIC AIC value

call the inputs for the function

Methods print, summary, coef and AIC can be used to extract specific outputs.

Details

$$a,b > 0$$ $$x = 0,1,2,...$$ $$freq \ge 0$$

NOTE : If input parameters are not in given domain conditions necessary error messages will be provided to go further.

References

Young-Xu, Y. & Chan, K.A., 2008. Pooling overdispersed binomial data to estimate event rate. BMC medical research methodology, 8(1), p.58.

Available at: http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=2538541&tool=pmcentrez&rendertype=abstract.

Trenkler, G., 1996. Continuous univariate distributions. Computational Statistics & Data Analysis, 21(1), p.119.

Available at: http://linkinghub.elsevier.com/retrieve/pii/0167947396900158.

Hughes, G., 1993. Using the Beta-Binomial Distribution to Describe Aggregated Patterns of Disease Incidence. Phytopathology, 83(9), p.759.

Available at: http://www.apsnet.org/publications/phytopathology/backissues/Documents/1993Abstracts/Phyto_83_759.htm

See also

Examples

No.D.D=0:7        #assigning the random variables
Obs.fre.1=c(47,54,43,40,40,41,39,95)   #assigning the corresponding frequencies

#estimating the parameters using maximum log likelihood value and assigning it
parameters=suppressWarnings(bbmle::mle2(EstMLEBetaBin,start = list(a=0.1,b=0.1),
data = list(x=No.D.D,freq=Obs.fre.1)))

bbmle::coef(parameters)   #extracting the parameters

#>         a         b 
#> 0.7229420 0.5808483 

#estimating the parameters using moment generating function methods
results<-EstMGFBetaBin(No.D.D,Obs.fre.1)

# extract the estimated parameters and summary
coef(results)

#>  	  a 	  b 
#>  0.7161628 0.5963324

summary(results)

#> Coefficients: 
#>  a 	  b 
#>  0.7161628 0.5963324
#> 
#> Negative Log-likelihood :  813.5872
#> 
#> AIC :  1631.174

AIC(results) #show the AIC value

#> [1] 1631.174