Exam Data

In an examination, there were 9 questions set on a particular topic. Each question is marked out of a total of 20 and in assessing the final class of a candidate, particular attention is paid to the total number of questions for which he has an "alpha", i.e., at least 15 out of 20, as well as his total number of marks. His number of alpha's is a rough indication of the "quality" of his exam performance. Thus, the distribution of alpha's over the candidates is of interest. There were 209 candidates attempting questions from this section of 9 questions and a total of 326 alpha's was awarded. So we treat 9 as the "litter size", and the dichotomous response is whether or not he got an alpha on the question.

Exam_data

Format

A data frame with 2 columns and 10 rows

No.of.alpha

No of Alphas

fre

Observed frequencies

Source

Extracted from

Paul, S.R., 1985. A three-parameter generalization of the binomial distribution. Communications in Statistics - Theory and Methods, 14(6), pp.1497-1506.

Available at: http://www.tandfonline.com/doi/abs/10.1080/03610928508828990

Examples

Exam_data$No.of.alpha              #extracting the binomial random variables

#>  [1] 0 1 2 3 4 5 6 7 8 9

sum(Exam_data$fre)                 #summing all the frequencies

#> [1] 209