Negative Log Likelihood value of Gamma Binomial Distribution

This function will calculate the Negative Log Likelihood value when the vector of binomial random variables and vector of corresponding frequencies are given with the shape parameters l and c.

Usage

NegLLGammaBin(x,freq,c,l)

Arguments

x

vector of binomial random variables.

freq

vector of frequencies.

c

single value for shape parameter c.

l

single value for shape parameter l.

Value

The output of NegLLGammaBin will produce a single numeric value.

Details

$$0 < l,c$$ $$freq \ge 0$$ $$x = 0,1,2,...$$

NOTE : If input parameters are not in given domain conditions necessary error messages will be provided to go further.

References

Grassia A (1977). “On a family of distributions with argument between 0 and 1 obtained by transformation of the gamma and derived compound distributions.” Australian Journal of Statistics, 19(2), 108--114.

Examples

No.D.D <- 0:7       #assigning the random variables
Obs.fre.1 <- c(47,54,43,40,40,41,39,95)   #assigning the corresponding frequencies

NegLLGammaBin(No.D.D,Obs.fre.1,.3,.4)   #acquiring the negative log likelihood value
#> [1] 825.7386