Negative Log Likelihood value of Kumaraswamy Binomial Distribution

This function will calculate the Negative Log Likelihood value when the vector of binomial random variables and vector of corresponding frequencies are given with the shape parameters a and b and iterations it.

NegLLKumBin(x,freq,a,b,it=25000)

Arguments

x
vector of binomial random variables.

freq
vector of frequencies.

a
single value for shape parameter alpha representing as a.

b
single value for shape parameter beta representing as b.

it
number of iterations to converge as a proper probability function replacing infinity.

Value

The output of NegLLKumBin will produce a single numeric value.

Details

$$0 < a,b $$ $$x = 0,1,2,...$$ $$freq \ge 0 $$ $$it > 0$$

NOTE : If input parameters are not in given domain conditions necessary error messages will be provided to go further.

References

Li, X. H., Huang, Y. Y., & Zhao, X. Y. (2011). The Kumaraswamy Binomial Distribution. Chinese Journal of Applied Probability and Statistics, 27(5), 511-521.

Examples

No.D.D <- 0:7          #assigning the random variables
Obs.fre.1 <- c(47,54,43,40,40,41,39,95)  #assigning the corresponding frequencies
# NOT RUN {
NegLLKumBin(No.D.D,Obs.fre.1,1.3,4.4) #acquiring the negative log likelihood value

# }