Overdispersion

After fitting the distribution using this function we can extract the overdispersion value. This function works for fitTriBin, fitBetaBin, fitKumBin, fitGHGBB and fitMcGBB for Binomial Mixture Distributions. Similarly, Alternate Binomial Distributions also support this function for fitAddBin,fitBetaCorrBin, fitCOMPBin, fitCorrBin and fitMultiBin.

Overdispersion(object)

Arguments

object
An object from one of the classes of fitTB,fitBB,fitKB,fitGB,fitMB.

Value

The output of Overdispersion gives a single value which is the overdispersion.

Details

Note : Only objects from classes of above mentioned classes can be used.

Examples

No.D.D=0:7      #assigning the random variables
Obs.fre.1=c(47,54,43,40,40,41,39,95)  #assigning the corresponding frequencies

#estimating mode value for given data
results<-EstMLETriBin(No.D.D,Obs.fre.1)
results

#> Call: 
#> EstMLETriBin(x = No.D.D, freq = Obs.fre.1)
#> 
#> Coefficient :
#>     mode 
#> 0.944444 

mode<-results$mode

#fitting the Triangular Bionomial distribution for estimated parameters
TriBin<-fitTriBin(No.D.D,Obs.fre.1,mode)
TriBin

#> Call: 
#> fitTriBin(x = No.D.D, obs.freq = Obs.fre.1, mode = mode)
#> 
#> Chi-squared test for Triangular Binomial Distribution 
#> 	
#>       Observed Frequency :  47 54 43 40 40 41 39 95 
#> 	
#>       expected Frequency :  11.74 23.47 35.21 46.94 58.66 70.2 79.57 73.21 
#> 	
#>       estimated Mode value: 0.944444 
#> 	
#>       X-squared : 193.6159   ,df : 6   ,p-value : 0 
#> 	
#>       over dispersion : 0.2308269 

#extracting the overdispersion
Overdispersion(TriBin)

#> [1] 0.2308269