Fitting the Additive Binomial Distribution when binomial random variable, frequency, probability of success and alpha are given

The function will fit the Additive Binomial distribution when random variables, corresponding frequencies, probability of success and alpha are given. It will provide the expected frequencies, chi-squared test statistics value, p value, and degree of freedom value so that it can be seen if this distribution fits the data.

fitAddBin(x,obs.freq,p,alpha)

Arguments

x
vector of binomial random variables.

obs.freq
vector of frequencies.

p
single value for probability of success.

alpha
single value for alpha.

Value

The output of fitAddBin gives the class format fitAB and fit consisting a list

bin.ran.var binomial random variables.

obs.freq corresponding observed frequencies.

exp.freq corresponding expected frequencies.

statistic chi-squared test statistics.

df degree of freedom.

p.value probability value by chi-squared test statistic.

fitAB fitted probability values of dAddBin.

NegLL Negative Log Likelihood value.

p estimated probability value.

alpha estimated alpha parameter value.

AIC AIC value.

call the inputs of the function.

Methods summary, print, AIC, residuals and fitted can be used to extract specific outputs.

Details

$$obs.freq \ge 0$$ $$x = 0,1,2,..$$ $$0 < p < 1$$ $$-1 < alpha < 1$$

References

Johnson, N. L., Kemp, A. W., & Kotz, S. (2005). Univariate discrete distributions (Vol. 444). Hoboken, NJ: Wiley-Interscience.

L. L. Kupper, J.K.H., 1978. The Use of a Correlated Binomial Model for the Analysis of Certain Toxicological Experiments. Biometrics, 34(1), pp.69-76.

Paul, S.R., 1985. A three-parameter generalization of the binomial distribution. Communications in Statistics - Theory and Methods, 14(6), pp.1497-1506.

Available at: http://www.tandfonline.com/doi/abs/10.1080/03610928508828990 .

Jorge G. Morel and Nagaraj K. Neerchal. Overdispersion Models in SAS. SAS Institute, 2012.

Examples

No.D.D=0:7         #assigning the random variables
Obs.fre.1=c(47,54,43,40,40,41,39,95)            #assigning the corresponding the frequencies

#assigning the estimated probability value

paddbin=suppressWarnings(EstMLEAddBin(No.D.D,Obs.fre.1)$p)

#assigning the estimated alpha value
alphaaddbin=suppressWarnings(EstMLEAddBin(No.D.D,Obs.fre.1)$alpha)

#fitting when the random variable,frequencies,probability and alpha are given
results<-fitAddBin(No.D.D,Obs.fre.1,paddbin,alphaaddbin)
results

#> Call: 
#> fitAddBin(x = No.D.D, obs.freq = Obs.fre.1, p = paddbin, alpha = alphaaddbin)
#> 
#> Chi-squared test for Additive Binomial Distribution 
#> 	
#>       Observed Frequency :  47 54 43 40 40 41 39 95 
#> 	
#>       expected Frequency :  10.71 50.11 79.86 45.82 24.86 74.29 84.07 29.28 
#> 	
#>       estimated p value : 0.546923  ,estimated alpha parameter : 0.230644 
#> 	
#>       X-squared : 336.831   ,df : 5   ,p-value : 0 

#extracting the AIC value
AIC(results)

#> [1] 930.6631

#extract fitted values
fitted(results)

#> [1] 10.71 50.11 79.86 45.82 24.86 74.29 84.07 29.28