Fitting the Beta-Correlated Binomial Distribution when binomial random variable, frequency, covariance, alpha and beta parameters are given

The function will fit the Beta-Correlated Binomial Distribution when random variables, corresponding frequencies, covariance, alpha and beta parameters are given. It will provide the expected frequencies, chi-squared test statistics value, p value, and degree of freedom so that it can be seen if this distribution fits the data.

fitBetaCorrBin(x,obs.freq,cov,a,b)

Arguments

x
vector of binomial random variables.

obs.freq
vector of frequencies.

cov
single value for covariance.

a
single value for alpha parameter.

b
single value for beta parameter.

Value

The output of fitBetaCorrBin gives the class format fitBCB and fit consisting a list

bin.ran.var binomial random variables.

obs.freq corresponding observed frequencies.

exp.freq corresponding expected frequencies.

statistic chi-squared test statistics.

df degree of freedom.

p.value probability value by chi-squared test statistic

corr Correlation value.

fitBCB fitted probability values of dBetaCorrBin.

NegLL Negative Log Likelihood value.

a estimated shape parameter value a.

b estimated shape parameter value b.

cov estimated covariance value.

AIC AIC value.

call the inputs of the function.

Methods summary, print, AIC, residuals and fitted can be used to extract specific outputs.

Details

$$obs.freq \ge 0$$ $$x = 0,1,2,..$$ $$-\infty < cov < +\infty$$ $$0 < a,b$$

NOTE : If input parameters are not in given domain conditions necessary error messages will be provided to go further.

References

Paul, S.R., 1985. A three-parameter generalization of the binomial distribution. Communications in Statistics - Theory and Methods, 14(6), pp.1497-1506.

Available at: http://www.tandfonline.com/doi/abs/10.1080/03610928508828990 .

Examples

No.D.D=0:7                    #assigning the random variables
Obs.fre.1=c(47,54,43,40,40,41,39,95)      #assigning the corresponding frequencies

#estimating the parameters using maximum log likelihood value and assigning it
parameters=suppressWarnings(bbmle::mle2(EstMLEBetaCorrBin,start = list(cov=0.0050,a=10,b=10),
           data = list(x=No.D.D,freq=Obs.fre.1)))

covBetaCorrBin=bbmle::coef(parameters)[1]
aBetaCorrBin=bbmle::coef(parameters)[2]
bBetaCorrBin=bbmle::coef(parameters)[3]

#fitting when the random variable,frequencies,covariance, a and b are given
results<-fitBetaCorrBin(No.D.D,Obs.fre.1,covBetaCorrBin,aBetaCorrBin,bBetaCorrBin)
results

#> Call: 
#> fitBetaCorrBin(x = No.D.D, obs.freq = Obs.fre.1, cov = covBetaCorrBin, 
#>     a = aBetaCorrBin, b = bBetaCorrBin)
#> 
#> Chi-squared test for Beta-Correlated Binomial Distribution 
#> 	
#>       Observed Frequency :  47 54 43 40 40 41 39 95 
#> 	
#>       expected Frequency :  48.71 47.41 45.42 43.16 39.81 36.91 42.44 95.14 
#> 	
#>       estimated covariance value: 0.07068406 
#> 	
#>       estimated a parameter : 3.199448  , estimated b parameter : 2.632928 
#> 	
#>       X-squared : 2.0695   ,df : 4   ,p-value : 0.723 

#extract AIC value
AIC(results)

#>        a 
#> 815.6536 

#extract fitted values
fitted(results)

#> [1] 48.71 47.41 45.42 43.16 39.81 36.91 42.44 95.14