Fitting the Binomial Distribution when binomial random variable, frequency and probability value are given

The function will fit the Binomial distribution when random variables, corresponding frequencies and probability value are given. It will provide the expected frequencies, chi-squared test statistics value, p value and degree of freedom so that it can be seen if this distribution fits the data.

fitBin(x,obs.freq,p=0)

Arguments

x
vector of binomial random variables.

obs.freq
vector of frequencies.

p
single value for probability.

Value

The output of fitBin gives the class format fitB and fit consisting a list

bin.ran.var binomial random variables.

obs.freq corresponding observed frequencies.

exp.freq corresponding expected frequencies.

statistic chi-squared test statistics value.

df degree of freedom.

p.value probability value by chi-squared test statistic.

fitB fitted probability values of dbinom.

phat estimated probability value.

call the inputs of the function.

Details

$$x = 0,1,2,...$$ $$0 \le p \le 1$$ $$obs.freq \ge 0$$

NOTE : If input parameters are not in given domain conditions necessary error messages will be provided to go further.

Examples

No.D.D <- 0:7      #assigning the random variables
Obs.fre.1 <- c(47,54,43,40,40,41,39,95)  #assigning the corresponding frequencies

#fitting when the random variable,frequencies are given.
fitBin(No.D.D,Obs.fre.1)

#> Chi-squared approximation may be doubtful because expected frequency is less than 5

#> Call: 
#> fitBin(x = No.D.D, obs.freq = Obs.fre.1)
#> 
#> Chi-squared test for Binomial Distribution 
#> 	
#>       Observed Frequency :  47 54 43 40 40 41 39 95 
#> 	
#>       expected Frequency :  1.59 13.41 48.3 96.68 116.11 83.66 33.49 5.75 
#> 	
#>       estimated probability value : 0.5456498 
#> 	
#>       X-squared : 2911.434   ,df : 6   ,p-value : 0