Fitting the Correlated Binomial Distribution when binomial random variable, frequency, probability of success and covariance are given

The function will fit the Correlated Binomial Distribution when random variables, corresponding frequencies, probability of success and covariance are given. It will provide the expected frequencies, chi-squared test statistics value, p value, and degree of freedom so that it can be seen if this distribution fits the data.

fitCorrBin(x,obs.freq,p,cov)

Arguments

x
vector of binomial random variables.

obs.freq
vector of frequencies.

p
single value for probability of success.

cov
single value for covariance.

Value

The output of fitCorrBin gives the class format fitCB and fit consisting a list

bin.ran.var binomial random variables.

obs.freq corresponding observed frequencies.

exp.freq corresponding expected frequencies.

statistic chi-squared test statistics.

df degree of freedom.

p.value probability value by chi-squared test statistic.

corr Correlation value.

fitCB fitted probability values of dCorrBin.

NegLL Negative Log Likelihood value.

AIC AIC value.

call the inputs of the function.

Methods summary, print, AIC, residuals and fitted can be used to extract specific outputs.

Details

$$obs.freq \ge 0$$ $$x = 0,1,2,..$$ $$0 < p < 1$$ $$-\infty < cov < +\infty$$

NOTE : If input parameters are not in given domain conditions necessary error messages will be provided to go further.

References

Johnson, N. L., Kemp, A. W., & Kotz, S. (2005). Univariate discrete distributions (Vol. 444). Hoboken, NJ: Wiley-Interscience.

L. L. Kupper, J.K.H., 1978. The Use of a Correlated Binomial Model for the Analysis of Certain Toxicological Experiments. Biometrics, 34(1), pp.69-76.

Paul, S.R., 1985. A three-parameter generalization of the binomial distribution. Communications in Statistics - Theory and Methods, 14(6), pp.1497-1506.

Available at: http://www.tandfonline.com/doi/abs/10.1080/03610928508828990.

Jorge G. Morel and Nagaraj K. Neerchal. Overdispersion Models in SAS. SAS Institute, 2012.

Examples

No.D.D=0:7       #assigning the random variables
Obs.fre.1=c(47,54,43,40,40,41,39,95)      #assigning the corresponding frequencies

#estimating the parameters using maximum log likelihood value and assigning it
parameters=suppressWarnings(bbmle::mle2(EstMLECorrBin,start = list(p=0.5,cov=0.0050),
           data = list(x=No.D.D,freq=Obs.fre.1)))

pCorrBin=bbmle::coef(parameters)[1]
covCorrBin=bbmle::coef(parameters)[2]

#fitting when the random variable,frequencies,probability and covariance are given
results<-fitCorrBin(No.D.D,Obs.fre.1,pCorrBin,covCorrBin)
results

#> Call: 
#> fitCorrBin(x = No.D.D, obs.freq = Obs.fre.1, p = pCorrBin, cov = covCorrBin)
#> 
#> Chi-squared test for Correlated Binomial Distribution 
#> 	
#>       Observed Frequency :  47 54 43 40 40 41 39 95 
#> 	
#>       expected Frequency :  10.7 50.1 79.85 45.84 24.87 74.29 84.07 29.28 
#> 	
#>       estimated p value : 0.5469539  ,estimated cov value : 0.05714648 
#> 	
#>       X-squared : 336.9971   ,df : 5   ,p-value : 0 

#extracting the AIC value
AIC(results)

#>        p 
#> 930.6631 

#extract fitted values
fitted(results)

#> [1] 10.70 50.10 79.85 45.84 24.87 74.29 84.07 29.28