Fitting the Gaussian Hypergeometric Generalized Beta Binomial Distribution when binomial random variable, frequency and shape parameters a,b and c are given

The function will fit the Gaussian Hypergeometric Generalized Beta Binomial Distribution when random variables, corresponding frequencies and shape parameters are given. It will provide the expected frequencies, chi-squared test statistics value, p value, degree of freedom and over dispersion value so that it can be seen if this distribution fits the data.

fitGHGBB(x,obs.freq,a,b,c)

Arguments

x
vector of binomial random variables.

obs.freq
vector of frequencies.

a
single value for shape parameter alpha representing a.

b
single value for shape parameter beta representing b.

c
single value for shape parameter lambda representing c.

Value

The output of fitGHGBB gives the class format fitGB and fit consisting a list

bin.ran.var binomial random variables.

obs.freq corresponding observed frequencies.

exp.freq corresponding expected frequencies.

statistic chi-squared test statistics.

df degree of freedom.

p.value probability value by chi-squared test statistic.

fitGB fitted values of dGHGBB.

NegLL Negative Loglikelihood value.

a estimated value for alpha parameter as a.

b estimated value for beta parameter as b.

c estimated value for gamma parameter as c.

AIC AIC value.

over.dis.para over dispersion value.

call the inputs of the function.

Methods summary, print, AIC, residuals and fitted can be used to extract specific outputs.

Details

$$0 < a,b,c$$ $$x = 0,1,2,...$$ $$obs.freq \ge 0$$

NOTE : If input parameters are not in given domain conditions necessary error messages will be provided to go further.

References

Rodriguez-Avi, J., Conde-Sanchez, A., Saez-Castillo, A. J., & Olmo-Jimenez, M. J. (2007). A generalization of the beta-binomial distribution. Journal of the Royal Statistical Society. Series C (Applied Statistics), 56(1), 51-61.

Available at : http://dx.doi.org/10.1111/j.1467-9876.2007.00564.x

Pearson, J., 2009. Computation of Hypergeometric Functions. Transformation, (September), p.1--123.

See also

Examples

No.D.D=0:7        #assigning the random variables
Obs.fre.1=c(47,54,43,40,40,41,39,95)       #assigning the corresponding frequencies

#estimating the parameters using maximum log likelihood value and assigning it
parameters=suppressWarnings(bbmle::mle2(EstMLEGHGBB,start = list(a=0.1,b=0.1,c=0.2),
data = list(x=No.D.D,freq=Obs.fre.1)))

bbmle::coef(parameters)         #extracting the parameters

#>         a         b         c 
#> 1.3508453 0.3245087 0.7004847 

aGHGBB=bbmle::coef(parameters)[1]  #assigning the estimated a
bGHGBB=bbmle::coef(parameters)[2]  #assigning the estimated b
cGHGBB=bbmle::coef(parameters)[3]  #assigning the estimated c

#fitting when the random variable,frequencies,shape parameter values are given.
results<-fitGHGBB(No.D.D,Obs.fre.1,aGHGBB,bGHGBB,cGHGBB)
results

#> Call: 
#> fitGHGBB(x = No.D.D, obs.freq = Obs.fre.1, a = aGHGBB, b = bGHGBB, 
#>     c = cGHGBB)
#> 
#> Chi-squared test for Gaussian Hypergeometric Generalized Beta-Binomial Distribution 
#> 	
#>       Observed Frequency :  47 54 43 40 40 41 39 95 
#> 	
#>       expected Frequency :  47.87 50.14 46.52 42.08 38.58 37.32 41.78 94.72 
#> 	
#>       estimated a parameter : 1.350845   ,estimated b parameter : 0.3245087  ,estimated c parameter : 0.7004847 
#> 	
#>       X-squared : 1.2831   ,df : 4   ,p-value : 0.8642 
#> 	
#>       over dispersion : 0.4324867 

#extracting the expected frequencies
fitted(results)

#> [1] 47.87 50.14 46.52 42.08 38.58 37.32 41.78 94.72

#extracting the residuals
residuals(results)

#> [1] -0.87  3.86 -3.52 -2.08  1.42  3.68 -2.78  0.28