Fitting the Grassia II Binomial distribution when binomial random variable, frequency and shape parameters are given

The function will fit the Grassia II Binomial Distribution when random variables, corresponding frequencies and shape parameters are given. It will provide the expected frequencies, chi-squared test statistics value, p value, degree of freedom and over dispersion value so that it can be seen if this distribution fits the data.

fitGrassiaIIBin(x,obs.freq,a,b)

Arguments

x
vector of binomial random variables.

obs.freq
vector of frequencies.

a
single value for shape parameter a.

b
single value for shape parameter b.

Value

The output of fitGrassiaIIBin gives the class format fitGrIIB and fit consisting a list

bin.ran.var binomial random variables.

obs.freq corresponding observed frequencies.

exp.freq corresponding expected frequencies.

statistic chi-squared test statistics.

df degree of freedom.

p.value probability value by chi-squared test statistic.

fitGrIIB fitted values of dGrassiaIIBin.

NegLL Negative Log Likelihood value.

a estimated value for shape parameter a.

b estimated value for shape parameter b.

AIC AIC value.

over.dis.para over dispersion value.

call the inputs of the function.

Methods summary, print, AIC, residuals and fitted can be used to extract specific outputs.

Details

$$0 < a,b$$ $$x = 0,1,2,...$$ $$obs.freq \ge 0$$

NOTE : If input parameters are not in given domain conditions necessary error messages will be provided to go further.

References

Grassia, A., 1977. On a family of distributions with argument between 0 and 1 obtained by transformation of the gamma and derived compound distributions. Australian Journal of Statistics, 19(2), pp.108-114.

Examples

No.D.D=0:7       #assigning the random variables
Obs.fre.1=c(47,54,43,40,40,41,39,95)          #assigning the corresponding frequencies

#estimating the parameters using maximum log likelihood value and assigning it
parameters=suppressWarnings(bbmle::mle2(EstMLEGrassiaIIBin,start = list(a=0.1,b=0.1),
data = list(x=No.D.D,freq=Obs.fre.1)))

aGIIBin=bbmle::coef(parameters)[1]         #assigning the estimated a
bGIIBin=bbmle::coef(parameters)[2]         #assigning the estimated b

#fitting when the random variable,frequencies,shape parameter values are given.
results<-fitGrassiaIIBin(No.D.D,Obs.fre.1,aGIIBin,bGIIBin)
results

#> Call: 
#> fitGrassiaIIBin(x = No.D.D, obs.freq = Obs.fre.1, a = aGIIBin, 
#>     b = bGIIBin)
#> 
#> Chi-squared test for Grassia II Binomial Distribution 
#> 	
#>       Observed Frequency :  47 54 43 40 40 41 39 95 
#> 	
#>       expected Frequency :  55.02 42.36 39.08 38.51 39.78 43.39 52.13 88.74 
#> 	
#>       estimated a parameter : 0.7285048   ,estimated b parameter : 2.025147  
#> 	
#>       X-squared : 8.6999   ,df : 5   ,p-value : 0.1216 
#> 	
#>       over dispersion : 0.2590043 

#extracting the expected frequencies
fitted(results)

#> [1] 55.02 42.36 39.08 38.51 39.78 43.39 52.13 88.74

#extracting the residuals
residuals(results)

#> [1]  -8.02  11.64   3.92   1.49   0.22  -2.39 -13.13   6.26