Fitting the Kumaraswamy Binomial Distribution when binomial random variable, frequency and shape parameters a and b, iterations parameter it are given

The function will fit the Kumaraswamy Binomial distribution when random variables, corresponding frequencies and shape parameters are given. It will provide the expected frequencies, chi-squared test statistics value, p value, degree of freedom and over dispersion value so that it can be seen if this distribution fits the data.

fitKumBin(x,obs.freq,a,b,it)

Arguments

x
vector of binomial random variables.

obs.freq
vector of frequencies.

a
single value for shape parameter alpha representing a.

b
single value for shape parameter beta representing b.

it
number of iterations to converge as a proper probability function replacing infinity.

Value

The output of fitKumBin gives the class format fitKB and fit consisting a list

bin.ran.var binomial random variables.

obs.freq corresponding observed frequencies.

exp.freq corresponding expected frequencies.

statistic chi-squared test statistics.

df degree of freedom.

p.value probability value by chi-squared test statistic.

fitKB fitted values of dKumBin.

NegLL Negative Log Likelihood value.

a estimated value for alpha parameter as a.

b estimated value for beta parameter as b.

it estimated it value for iterations.

AIC AIC value.

over.dis.para over dispersion value.

call the inputs of the function.

Methods summary, print, AIC, residuals and fiited can be used to extract specific outputs.

Details

$$0 < a,b$$ $$x = 0,1,2,...n$$ $$obs.freq \ge 0$$ $$it > 0$$

NOTE : If input parameters are not in given domain conditions necessary error messages will be provided to go further.

References

Li, X. H., Huang, Y. Y., & Zhao, X. Y. (2011). The Kumaraswamy Binomial Distribution. Chinese Journal of Applied Probability and Statistics, 27(5), 511-521.

See also

Examples

No.D.D=0:7   #assigning the random variables
Obs.fre.1=c(47,54,43,40,40,41,39,95)   #assigning the corresponding frequencies

#estimating the parameters using maximum log likelihood value and assigning it

parameters=suppressWarnings(bbmle::mle2(EstMLEKumBin,start = list(a=10.1,b=1.1,it=10000),
          data = list(x=No.D.D,freq=Obs.fre.1)))

bbmle::coef(parameters)    #extracting the parameters

#>            a            b           it 
#>    0.7302994    0.6162630 9999.9957297 

aKumBin=bbmle::coef(parameters)[1] #assigning the estimated a
bKumBin=bbmle::coef(parameters)[2] #assigning the estimated b
itKumBin=bbmle::coef(parameters)[3] #assigning the estimated iterations

#fitting when the random variable,frequencies,shape parameter values are given.
results<-fitKumBin(No.D.D,Obs.fre.1,aKumBin,bKumBin,itKumBin*100)
results

#> Call: 
#> fitKumBin(x = No.D.D, obs.freq = Obs.fre.1, a = aKumBin, b = bKumBin, 
#>     it = itKumBin * 100)
#> 
#> Chi-squared test for Kumaraswamy Binomial Distribution 
#> 	
#>       Observed Frequency :  47 54 43 40 40 41 39 95 
#> 	
#>       expected Frequency :  53.33 42.16 39.59 39.52 41.19 45.09 53.71 84.37 
#> 	
#>       estimated a parameter : 0.7302994   ,estimated b parameter : 0.616263  ,estimated it value : 999999.6 
#> 	
#>       X-squared : 10.1494   ,df : 5   ,p-value : 0.0711 
#> 	
#>       over dispersion : 0.4230119 

#extracting the expected frequencies
fitted(results)

#> [1] 53.33 42.16 39.59 39.52 41.19 45.09 53.71 84.37

#extracting the residuals
residuals(results)

#> [1]  -6.33  11.84   3.41   0.48  -1.19  -4.09 -14.71  10.63