Fitting the Lovinson Multiplicative Binomial Distribution when binomial random variable, frequency, probability of success and theta parameter are given

The function will fit the Lovinson Multiplicative Binomial distribution when random variables, corresponding frequencies, probability of success and phi parameter are given. It will provide the expected frequencies, chi-squared test statistics value, p value and degree of freedom value so that it can be seen if this distribution fits the data.

fitLMBin(x,obs.freq,p,phi)

Arguments

x
vector of binomial random variables.

obs.freq
vector of frequencies.

p
single value for probability of success.

phi
single value for phi parameter.

Value

The output of fitLMBin gives the class format fitLMB and fit consisting a list

bin.ran.var binomial random variables.

obs.freq corresponding observed frequencies.

exp.freq corresponding expected frequencies.

statistic chi-squared test statistics.

df degree of freedom.

p.value probability value by chi-squared test statistic.

fitLMB fitted probability values of dLMBin.

NegLL Negative Log Likelihood value.

p estimated probability value.

phi estimated phi parameter value.

AIC AIC value.

call the inputs of the function.

Methods summary, print, AIC, residuals and fitted can be used to extract specific outputs.

Details

$$obs.freq \ge 0$$ $$x = 0,1,2,..$$ $$0 < p < 1$$ $$0 < phi $$

References

Elamir, E.A., 2013. Multiplicative-Binomial Distribution: Some Results on Characterization, Inference and Random Data Generation. Journal of Statistical Theory and Applications, 12(1), pp.92-105.

See also

Examples

No.D.D=0:7       #assigning the random variables
Obs.fre.1=c(47,54,43,40,40,41,39,95)     #assigning the corresponding frequencies

#estimating the parameters using maximum log likelihood value and assigning it
parameters=suppressWarnings(bbmle::mle2(EstMLELMBin,start = list(p=0.1,phi=.3),
          data = list(x=No.D.D,freq=Obs.fre.1)))

pLMBin=bbmle::coef(parameters)[1]    #assigning the estimated probability value
phiLMBin=bbmle::coef(parameters)[2]  #assigning the estimated phi value

#fitting when the random variable,frequencies,probability and phi are given
results<-fitLMBin(No.D.D,Obs.fre.1,pLMBin,phiLMBin)
results

#> Call: 
#> fitLMBin(x = No.D.D, obs.freq = Obs.fre.1, p = pLMBin, phi = phiLMBin)
#> 
#> Chi-squared test for Lovinson Multiplicative Binomial Distribution 
#> 	
#>       Observed Frequency :  47 54 43 40 40 41 39 95 
#> 	
#>       expected Frequency :  54.3 49.54 38.86 33.97 35.74 45.26 63.86 77.48 
#> 	
#>       estimated p value : 0.5126962  ,estimated phi parameter : 0.7060546 
#> 	
#>       X-squared : 17.4425   ,df : 5   ,p-value : 0.0037 

#extracting the AIC value
AIC(results)

#>        p 
#> 821.9933 

#extract fitted values
fitted(results)

#> [1] 54.30 49.54 38.86 33.97 35.74 45.26 63.86 77.48