Binomial Data Extraction from Raw data

The below function has the ability to extract from the raw data to Binomial Outcome Data. This function simplifies the data into more presentable way to the user.

BODextract(data)

Arguments

data
vector of observations

Value

The output of BODextract gives a list format consisting

RV binomial random variables in vector form

Freq corresponding frequencies in vector form

Details

NOTE : If input parameters are not in given domain conditions necessary error messages will be provided to go further

Examples

datapoints=sample(0:10,340,replace=TRUE) #creating a sample set of observations
BODextract(datapoints)                   #extracting binomial outcome data from observations

#> $RV
#>  [1]  0  1  2  3  4  5  6  7  8  9 10
#> 
#> $Freq
#>  [1] 33 36 29 26 33 34 32 26 31 36 24
#> 

Random.variable=BODextract(datapoints)$RV #extracting the binomial random variables